Also use mocked config in test_lisflood
